refactor: Remove use of 'long' integral type.
The 'long' type varies in size between different 64bit ABIs and as such is an issue for portability.

+++ Code/BinkMovie/BinkPlayer.h
@@ -34,7 +34,7 @@ class BINKMovieClass : public MovieClass
 		HBINK Bink;
 		bool FrameChanged;
 		unsigned TextureCount;
-		unsigned long TicksPerFrame;
+		unsigned int TicksPerFrame;
 
 		struct TextureInfoStruct {
 			TextureClass* Texture;
