+++ Code/Combat/damage.cpp
@@ -1134,7 +1134,7 @@ void DefenseObjectClass::Export(BitStreamClass & packet)
 
 	packet.Add(health,							BITPACK_HEALTH);
 	packet.Add(shield_strength,				BITPACK_SHIELD_STRENGTH);
-	packet.Add((unsigned long)ShieldType,	BITPACK_SHIELD_TYPE);
+	packet.Add((unsigned int)ShieldType,	BITPACK_SHIELD_TYPE);
 
    //LastSentHealth = Health;
    //LastSentSkin = Skin;
