+++ Code/Commando/gamesideservercontrol.h
@@ -63,7 +63,7 @@ class GameSideServerControlClass
 		static void Shutdown(void);
 		static const char *App_Request_Callback(char *request);
 		static void Print(const char *text, ...);
-		static void Send_Message(const char *text, unsigned long ip, unsigned short port);
+		static void Send_Message(const char *text, unsigned int ip, unsigned short port);
 		static void Set_Welcome_Message(void);
 
 	private:
